feat: add edge specific code
