+++ packages/core/src/readers/index.edge-light.ts
@@ -1,0 +1,3 @@
+export * from "./SimpleDirectoryReader.edge-light.js";
+export * from "./TextFileReader.js";
+export * from "./type.js";
