+++ packages/core/src/readers/index.ts
@@ -9,4 +9,5 @@ export * from "./NotionReader.js";
 export * from "./PDFReader.js";
 export * from "./SimpleDirectoryReader.js";
 export * from "./SimpleMongoReader.js";
+export * from "./TextFileReader.js";
 export * from "./type.js";
