+++ packages/core/src/storage/index.edge.ts
@@ -1,0 +1,13 @@
+export * from "./FileSystem.js";
+export * from "./StorageContext.js";
+export { SimpleChatStore } from "./chatStore/SimpleChatStore.js";
+export * from "./chatStore/types.js";
+export * from "./constants.js";
+export { SimpleDocumentStore } from "./docStore/SimpleDocumentStore.js";
+export * from "./docStore/types.js";
+export { SimpleIndexStore } from "./indexStore/SimpleIndexStore.js";
+export * from "./indexStore/types.js";
+export { SimpleKVStore } from "./kvStore/SimpleKVStore.js";
+export * from "./kvStore/types.js";
+export { SimpleVectorStore } from "./vectorStore/SimpleVectorStore.js";
+export * from "./vectorStore/types.js";
