+++ packages/core/src/storage/index.ts
@@ -1,14 +1,4 @@
-export * from "./FileSystem.js";
-export * from "./StorageContext.js";
-export { SimpleChatStore } from "./chatStore/SimpleChatStore.js";
-export * from "./chatStore/types.js";
-export * from "./constants.js";
-export { SimpleDocumentStore } from "./docStore/SimpleDocumentStore.js";
-export * from "./docStore/types.js";
-export { SimpleIndexStore } from "./indexStore/SimpleIndexStore.js";
-export * from "./indexStore/types.js";
-export { SimpleKVStore } from "./kvStore/SimpleKVStore.js";
-export * from "./kvStore/types.js";
+export * from "./index.edge.js";
 export { AstraDBVectorStore } from "./vectorStore/AstraDBVectorStore.js";
 export { ChromaVectorStore } from "./vectorStore/ChromaVectorStore.js";
 export { MilvusVectorStore } from "./vectorStore/MilvusVectorStore.js";
@@ -17,4 +7,3 @@ export { PGVectorStore } from "./vectorStore/PGVectorStore.js";
 export { PineconeVectorStore } from "./vectorStore/PineconeVectorStore.js";
 export { QdrantVectorStore } from "./vectorStore/QdrantVectorStore.js";
 export { SimpleVectorStore } from "./vectorStore/SimpleVectorStore.js";
-export * from "./vectorStore/types.js";
